+++ src/PbiIndexIO.h
@@ -39,6 +39,7 @@
 #define PBIINDEXIO_H
 
 #include "pbbam/BamFile.h"
+#include "pbbam/DataSet.h"
 #include "pbbam/PbiFile.h"
 #include "pbbam/PbiRawData.h"
 #include <htslib/bgzf.h>
@@ -57,10 +58,9 @@ class PbiIndexIO
 public:
     // top-level entry points
     static PbiRawData Load(const std::string& filename);
-    static void Load(PbiRawData& rawData,
-                     const std::string& filename);
-    static void Save(const PbiRawData& rawData,
-                     const std::string& filename);
+    static void Load(PbiRawData& rawData, const std::string& filename);
+    static void LoadFromDataSet(PbiRawData& aggregateData, const DataSet& dataset);
+    static void Save(const PbiRawData& rawData, const std::string& filename);
 
 public:
     // per-component load